% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_zip_sf}
\alias{nyc_zip_sf}
\title{Simple Features collection of NYC Zip Codes}
\format{
\subsection{\code{nyc_zip_sf}}{

A simple feature collection with 212 rows and 4 columns:
\describe{
\item{zip}{Zip code (actually ZTCA)}
\item{zip_name}{ZCTA5 name}
\item{pop}{2019 ACS population estimage}
\item{geometry}{Multipolydon for zip code area}
}
}
}
\source{
See details.
}
\usage{
nyc_zip_sf
}
\description{
Polygons of NYC zip codes along with 2019 ACS population estimate.
}
\details{
Geographies and population estimates from US Census Bureau, ACS 5-year estimates (2019).
List of zip codes described in \code{nyc_zips} table elsewhere in this package.
}
\author{
Kieran Healy
}
\keyword{datasets}
