% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_zips}
\alias{nyc_zips}
\title{NYC Zip Codes}
\format{
\subsection{\code{nyc_zips}}{

A data frame with 211 rows and 6 columns:
\describe{
\item{zip}{Zip Code}
\item{borough}{Borough name}
\item{city}{Nominal city (USPS designation)}
\item{county}{County as 'Countyname County'}
\item{long_county}{County as 'Countyname County, New York'}
\item{short_county}{County as 'County'}
}
}
}
\source{
See details.
}
\usage{
nyc_zips
}
\description{
Table of New York City zip codes.
}
\details{
Full USPS Zip Code list from \url{https://postalpro.usps.com/ZIP_Locale_Detail}
Initial regional subsetting via \url{https://worldpopulationreview.com/zips/new-york}
}
\author{
Kieran Healy
}
\keyword{datasets}
